#40
This fix remove SenderRule and ReceiptRule because they are always
problemetic. We should create two differents streams if ReceiptRule are
specific for a destination. It's hard to implement and we got a lot of
issues with this feature. Some parsing errors was just bypassed rules
and it's not a really usefull feature.

This fix also improve performance of the agent.